+++ samples/java/08objects/main/Shape.java
@@ -1,6 +1,6 @@
 package main;
 
 public abstract class Shape {
-    public abstract float area();
-    public abstract float perimeter();
+    public abstract double area();
+    public abstract double perimeter();
 }
